+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added support for implementing OAuth1 servers.
 - Implemented RSA-SHA1 signature method.
 - Fixed support for signing parameters with multiple values.
+- Fixed percent encoding to generate correct signature base strings.
 
 ## 1.0.3
 - Confirm compatibility with package http 0.12.0